Revert "Merge branch 'PlayerBehavior' into AudioManager"
This reverts commit 6d3b472913837a6d766313f253709a5343f973b6, reversing
changes made to 142aaec3947e996fb39dfc2d36a37e2c74888c19.
