--- Project Chronograph/Assets/Art/Animations/Animations/Player/Player_RedFlash.anim.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: ac79e791714b2446aae4a2a81c867d01
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 7400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
