--- Project Chronograph/Assets/Prefabs/RaycastPlayer.prefab.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 0f2e3a5d466e045d09140836537fba86
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 100100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
