--- Project Chronograph/Assets/Scripts/Environment Scripts/PlatformMover.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 5bdd85e847fbf44c1a787f67ab32203c
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
