+++ Project Chronograph/Assets/Scripts/Player Scripts/AttackTrigger.cs
@@ -5,13 +5,13 @@ using UnityEngine;
 public class AttackTrigger : MonoBehaviour {
 
     public int damage = 1;
-    public EnemyHealthManager enemyHealthMan;
+    public EnemyHealthManager Ehealthman;
 
     void OnTrggerEnter2D(Collider2D col){
 
         if(col.isTrigger != true && col.CompareTag("Enemy")){
 
-            enemyHealthMan.giveDamage(damage);
+            Ehealthman.giveDamage(damage);
 
         }
