--- Project Chronograph/Assets/Scripts/Player Scripts/Raycast Scripts.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 33034149a58444b62a3c5e7c91b2f6a2
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
