--- Project Chronograph/Assets/Scripts/Player Scripts/Raycast Scripts/RaycastCheckTouching.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: e93dcd6f356164af0a40db1fe62e778e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
