+++ Project Chronograph/ProjectSettings/UnityConnectSettings.asset
@@ -11,13 +11,13 @@ UnityConnectSettings:
   CrashReportingSettings:
     m_EventUrl: https://perf-events.cloud.unity3d.com/api/events/crashes
     m_NativeEventUrl: https://perf-events.cloud.unity3d.com/symbolicate
-    m_Enabled: 1
+    m_Enabled: 0
     m_CaptureEditorExceptions: 1
   UnityPurchasingSettings:
     m_Enabled: 0
     m_TestMode: 0
   UnityAnalyticsSettings:
-    m_Enabled: 0
+    m_Enabled: 1
     m_InitializeOnStartup: 1
     m_TestMode: 0
     m_TestEventUrl:
